+++ src/config/parameters.rs
@@ -1,9 +1,10 @@
-use super::fitness::FitnessModel;
 use serde::{Deserialize, Serialize};
 use std::fs;
 
+use crate::fitness::FitnessModel;
+
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct SimulationParameters {
+pub struct Parameters {
     pub mutation_rate: f64,
     pub recombination_rate: f64,
     pub host_population_size: usize,
@@ -28,12 +29,12 @@ pub struct FitnessModelFraction {
 }
 
 #[derive(Debug)]
-pub enum SimulationParametersError {
+pub enum ParametersError {
     IoError(std::io::Error),
     YamlError(serde_yaml::Error),
 }
 
-impl std::fmt::Display for SimulationParameters {
+impl std::fmt::Display for Parameters {
     fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut output = vec![];
         self.write(&mut output).map_err(|_| std::fmt::Error)?;
@@ -41,27 +42,23 @@ impl std::fmt::Display for SimulationParameters {
     }
 }
 
-impl SimulationParameters {
-    pub fn write(&self, writer: &mut dyn std::io::Write) -> Result<(), SimulationParametersError> {
-        serde_yaml::to_writer(writer, self).map_err(SimulationParametersError::YamlError)
+impl Parameters {
+    pub fn write(&self, writer: &mut dyn std::io::Write) -> Result<(), ParametersError> {
+        serde_yaml::to_writer(writer, self).map_err(ParametersError::YamlError)
     }
 
-    pub fn read(
-        reader: &mut dyn std::io::Read,
-    ) -> Result<SimulationParameters, SimulationParametersError> {
-        serde_yaml::from_reader(reader).map_err(SimulationParametersError::YamlError)
+    pub fn read(reader: &mut dyn std::io::Read) -> Result<Parameters, ParametersError> {
+        serde_yaml::from_reader(reader).map_err(ParametersError::YamlError)
     }
 
-    pub fn write_to_file(&self, filename: &str) -> Result<(), SimulationParametersError> {
-        let file = fs::File::create(filename).map_err(SimulationParametersError::IoError)?;
+    pub fn write_to_file(&self, filename: &str) -> Result<(), ParametersError> {
+        let file = fs::File::create(filename).map_err(ParametersError::IoError)?;
         let mut writer = std::io::BufWriter::new(file);
         self.write(&mut writer)
     }
 
-    pub fn read_from_file(
-        filename: &str,
-    ) -> Result<SimulationParameters, SimulationParametersError> {
-        let file = fs::File::open(filename).map_err(SimulationParametersError::IoError)?;
+    pub fn read_from_file(filename: &str) -> Result<Parameters, ParametersError> {
+        let file = fs::File::open(filename).map_err(ParametersError::IoError)?;
         let mut reader = std::io::BufReader::new(file);
         Self::read(&mut reader)
     }
@@ -75,7 +72,7 @@ mod tests {
     #[test]
     fn read_write_neutral() {
         let mut buffer = Vec::new();
-        let settings = SimulationParameters {
+        let settings = Parameters {
             mutation_rate: 1e-6,
             recombination_rate: 1e-8,
             substitution_matrix: [
@@ -95,14 +92,14 @@ mod tests {
             )),
         };
         settings.write(&mut buffer).unwrap();
-        let read_settings = SimulationParameters::read(&mut buffer.as_slice()).unwrap();
+        let read_settings = Parameters::read(&mut buffer.as_slice()).unwrap();
         assert_eq!(read_settings, settings);
     }
 
     #[test]
     fn read_write_exponential() {
         let mut buffer = Vec::new();
-        let settings = SimulationParameters {
+        let settings = Parameters {
             mutation_rate: 1e-6,
             recombination_rate: 1e-8,
             substitution_matrix: [
@@ -131,7 +128,7 @@ mod tests {
             )),
         };
         settings.write(&mut buffer).unwrap();
-        let read_settings = SimulationParameters::read(&mut buffer.as_slice()).unwrap();
+        let read_settings = Parameters::read(&mut buffer.as_slice()).unwrap();
         assert_eq!(read_settings, settings);
     }
 
@@ -139,7 +136,7 @@ mod tests {
     fn read_write_file() {
         let tmp_dir = std::env::temp_dir().join("test_settings.yaml");
         let path = tmp_dir.to_str().unwrap();
-        let settings = SimulationParameters {
+        let settings = Parameters {
             mutation_rate: 1e-6,
             recombination_rate: 1e-8,
             substitution_matrix: [
@@ -159,7 +156,7 @@ mod tests {
             )),
         };
         settings.write_to_file(path).unwrap();
-        let read_settings = SimulationParameters::read_from_file(path).unwrap();
+        let read_settings = Parameters::read_from_file(path).unwrap();
         assert_eq!(read_settings, settings);
         std::fs::remove_file(path).unwrap();
     }