+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 - Trees only report the `block_id`
 - Include `block_id` in fasta output
 - Truncate barcode file at the start of the simulation
+- Change naming in configuration files for consistency
+    - `SimulationSettings` -> `Settings`
+    - `SimulationParameters` -> `Parameters`
+    - `SimulationPlan` -> `Schedule`
+- Field names have been renamed accordingly
+- Moved all things related to `Settings` into config module
 
 ## 0.1.0 --- Rapid Expansion (Jul 26, 2023)